Add DTO project library
