+++ Farfetch.DTO/Boutique.cs
@@ -1,10 +1,9 @@
-﻿using System;
-namespace Farfetch
+﻿namespace Farfetch
 {
 	public class Boutique
 	{
 		public string ImageUri { get; set; }
 		public string Name { get; set; }
-		public string ShortAddress { get; set;}
+		public string ShortAddress { get; set; }
 	}
 }
