+++ Farfetch.DTO/Catalog.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Farfetch
+﻿namespace Farfetch
 {
 	public class Catalog
 	{
