+++ Farfetch.DTO/Item.cs
@@ -1,0 +1,11 @@
+﻿namespace Farfetch
+{
+	public class Item
+	{
+		public string Name { get; set; }
+		public decimal Price { get; set; }
+		public string ImageUri { get; set; }
+		public string Description { get; set; }
+		public bool IsStarred { get; set; }
+	}
+}
