+++ Farfetch.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Farfetch.Droid", "Farfetch\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Farfetch.iOS", "Farfetch\Farfetch.iOS\Farfetch.iOS.csproj", "{5F7CD482-6CE6-4413-9B5A-0F69ECDCDA13}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Farfetch.DTO", "Farfetch.DTO\Farfetch.DTO.csproj", "{744DCBBF-235B-447F-AB5C-022269C5D054}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Ad-Hoc|Any CPU = Ad-Hoc|Any CPU
@@ -106,6 +108,30 @@ Global
 		{5F7CD482-6CE6-4413-9B5A-0F69ECDCDA13}.Release|iPhoneSimulator.ActiveCfg = Release|iPhoneSimulator
 		{5F7CD482-6CE6-4413-9B5A-0F69ECDCDA13}.Release|iPhoneSimulator.Build.0 = Release|iPhoneSimulator
 		{5F7CD482-6CE6-4413-9B5A-0F69ECDCDA13}.Debug|Any CPU.Build.0 = Debug|iPhone
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|iPhone.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|iPhone.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Ad-Hoc|iPhoneSimulator.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|Any CPU.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|iPhone.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|iPhone.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.AppStore|iPhoneSimulator.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|Any CPU.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|iPhone.Build.0 = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{744DCBBF-235B-447F-AB5C-022269C5D054}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
